Find Catch2 via find_package
find_package is the canonical way to express dependencies in CMake
and conveniently provides a mechanism for override the default
package search logic by simply providing your own Find<Package>.cmake
file. This means when it comes time to use a system install of Catch2,
we can just delete this custom module.