Default to building a shared library
This is a workaround to deal with the fact that when RSL is packaged
for the ROS ecosystem, -DBUILD_SHARED_LIBS=ON is not specified. It's
disappointing that the packagers cannot just ask CMake for what they
want but alas this is required to ensure that they build shared libs
as they would prefer. Once OSRF learns how to build shared libraries
we can remove this.