Fix static_string.cpp compilation with MSVC standard library
std::array's iterator is a pointer in libc++ and libstdc++ so
`const auto*` works fine. However the array iterator is not a
pointer on MSVC so compilation fails. Removing the asterisk still
results in a pointer type being used but causes a clang-tidy check
to fail so we have to disable that.